% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rolling_zscore.R
\name{calculate_rolling_zscore}
\alias{calculate_rolling_zscore}
\title{Calculate Rolling Zscore}
\usage{
calculate_rolling_zscore(x, column, lookback = 20, column_name = zscore)
}
\arguments{
\item{x}{A data tibble}

\item{column}{The column variable to perform calculations}

\item{lookback}{The lookback period}

\item{column_name}{The Column's Output Name}
}
\value{
A tibble
}
\description{
Calculate Rolling Zscore
}
