% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliderFinance-package.R
\docType{package}
\name{sliderFinance-package}
\alias{sliderFinance}
\alias{sliderFinance-package}
\title{sliderFinance: A collection of Rolling Financial Function Using Slider}
\description{
A collection of Rolling Financial Functions using SLider Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
