% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rolling_market_regime.R
\name{calculate_rolling_market_regime}
\alias{calculate_rolling_market_regime}
\title{Calculate Rolling Market Regime}
\usage{
calculate_rolling_market_regime(
  x,
  column,
  lookback = 100,
  column_name = market_regime
)
}
\arguments{
\item{x}{A tibble}

\item{column}{The column variable to perform calculations}

\item{lookback}{The lookback period}

\item{column_name}{The Column's Output Name}
}
\value{
A tibble
}
\description{
Calculate Rolling Market Regime
}
