% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rolling_stochastic_oscillator.R
\name{calculate_rolling_stochastic_oscillator}
\alias{calculate_rolling_stochastic_oscillator}
\title{Calculate Rolling Stochastic Oscillator}
\usage{
calculate_rolling_stochastic_oscillator(x, column, lookback = 14)
}
\arguments{
\item{x}{A tibble}

\item{column}{The column variable to perform calculations}

\item{lookback}{The lookback period}
}
\value{
A tibble
}
\description{
Calculate Rolling Stochastic Oscillator
}
